move all types to top level in types package
