+++ packages/types/redux.js
@@ -1,1 +1,8 @@
-module.exports = require("./redux");
+// @flow
+/* eslint-disable no-undef */
+
+declare type Action = {
+  +type: string
+};
+
+declare type Dispatch<A: Action> = (action: A) => A;
