--- packages/types/src/channels.js
@@ -1,3 +1,0 @@
-// @flow
-
-export type Channels = rxjs$Subject<*>;
